fix: ERC7535 not properly calculating expectedShares
In ERC4626, when totalAssets() is called during the deposit function it calculates the total amount of assets PRIOR to deposit. In ERC7535, as the depositing asset is ETH, the ETH is already on the contract by the time this call is made and thus the total assets returns the amount of assets AFTER a deposit. Changes were made to account for this difference.