refactor: Move ERC7535 deposit fix from _convertToShares to deposit()
The fix for ensuring the users deposit is not included in a totalAssets call during a deposit function call was previously done in _convertToShares. While this resolved the issue, it meant that previewDeposit and convertToShares were now not functioning correctly as per the ERC4626 standard (not returning the amount a user would get on a function call at that block).