fix: Decimal issues on first deposit introduced by new OpenZeppelin ver
In OpenZeppelin v4.9, the decimals for the shares in an empty vault are now 1:1 with the asset, rather than being calculated and accounted for to ensure 18 decimals were returned. Essentially, we need to add in a fix to ensure that on empty vaults, if the asset is non 18 decimals, the vault shares should still be 18 decimals to ensure all vaults have same decimals.